requirements.txt
pandas >= 1.4
seaborn >= 0.11
streamlit >= 1.12.0
matplotlib >= 3.1
plotly >= 5.10.0
requests >= 2.31.0 
beautifulsoup4 >= 4.12.2 
mtranslate >= 1.8
nltk >= 3.8.1
geopy >= 2.4.0 
spacy >= 3.5.0
python-Levenshtein >= 0.21.1 
scikit-learn >= 1.3.0 
pydantic >= 1.8.2
wordcloud >= 1.9.2 
fr_core_news_sm-3.5.0-py3-none-any.whl
